Still trying to commit the change int->double for node width and height in applyVizmapSettings(). -Andrew 07-01-02
git-svn-id: http://chianti.ucsd.edu/svn@4933 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ cytoscape/trunk/src/cytoscape/CytoscapeWindow.java
@@ -374,9 +374,9 @@ protected void applyVizmapSettings ()
 	vizMapperCategories.getNodeFillColor(bundle, vizMapper);
     Color nodeBorderColor =
 	vizMapperCategories.getNodeBorderColor(bundle, vizMapper);
-    int nodeHeight =
+    double nodeHeight =
 	vizMapperCategories.getNodeHeight(bundle, vizMapper);
-    int nodeWidth =
+    double nodeWidth =
 	vizMapperCategories.getNodeWidth(bundle, vizMapper);
     byte nodeShape =
 	vizMapperCategories.getNodeShape(bundle, vizMapper);
